refactor: move Subdomain to its producer (factor_pairs)
Subdomain is constructed only in factor_pairs::split_into_subdomains and
is the return vocabulary of build_local_domains, yet it was defined in
the domain root — so factor_pairs reached *up* via super::Subdomain for a
type it produces itself. Move the struct next to its producer and
re-export it from domain.rs (as already done for build_local_domains).

Drops the factor_pairs -> domain-root edge for Subdomain and the root's
now-orphaned schwarz_precond::SubdomainCore import; factor_pairs already
depended on SubdomainCore. domain.rs keeps Design (its genuine core type)
and the consumer path crate::domain::Subdomain is unchanged.
